split away base_client from cosmos & evm mods
